more restructuring, erc20 should use Ownable instead of Capped, cleaner will not clean dbg.json
