+++ contracts/token/ERC1155/extensions/ERC1155Ownable.sol
@@ -3,7 +3,7 @@
 pragma solidity >=0.6.0;
 
 import "@openzeppelin/contracts/token/ERC1155/ERC1155.sol";
-import "../../utils/Ownable.sol";
+import "../../../access/Ownable.sol";
 
 contract ERC1155Ownable is ERC1155, Ownable {
