+++ contracts/token/ERC721/utils/Opener.sol
@@ -1,5 +1,8 @@
+// SPDX-License-Identifier: UNLICENSED
+pragma solidity >=0.6.0;
+
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
-import "../../../utils/Ownable.sol";
+import "../../../access/Ownable.sol";
 
 contract Opener is  Ownable {
     using SafeMath for uint256;
