+++ contracts/utils/StakingContract.sol
@@ -6,7 +6,7 @@ import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "@openzeppelin/contracts/utils/Pausable.sol";
 import "@openzeppelin/contracts/token/ERC721/IERC721.sol";
 import "@openzeppelin/contracts/math/SafeMath.sol";
-import "./Ownable.sol";
+import "../access/Ownable.sol";
 
 contract StakingContract is Pausable, Ownable {
     using SafeMath for uint256;
