+++ generate-typings.config.mjs
@@ -10,8 +10,7 @@ export default {
     /Network(Registry|V2)?/,
     /BountyToken/,
     /^ERC1155Ownable/,
-    /^ERC20(Ownable|Distribution|TokenLock)/,
-    /Token\.json$/,
+    /^ERC20(Ownable|Distribution|TokenLock|Capped)/,
     /^ERC4626/,
     /^ERC721(Collectibles|Standard)/,
     /StakingContract/,
