+++ test/utils/erc20-deployer.ts
@@ -8,14 +8,13 @@ import {ERC20, toSmartContractDecimals, Web3Connection, type Web3ConnectionOptio
  * @param web3Connection
  * @returns Promise<TransactionReceipt>
  */
-export async function erc20Deployer(name: string, symbol: string, cap = toSmartContractDecimals(1000000, 18), web3Connection: Web3Connection|Web3ConnectionOptions) {
+export async function erc20Deployer(name: string, symbol: string, cap = toSmartContractDecimals(1000000, 18), web3Connection: Web3Connection|Web3ConnectionOptions, decimals = 18) {
   if (!(web3Connection instanceof Web3Connection))
     web3Connection = new Web3Connection(web3Connection)
 
-  await web3Connection.start();
-  const address = await web3Connection.getAddress();
+  web3Connection.start();
 
   const deployer = new ERC20(web3Connection);
 
-  return deployer.deployJsonAbi(name, symbol, cap, address);
+  return deployer.deployJsonAbi(name, symbol, cap, decimals);
 }
